+++ api/database/models/examen.js
@@ -15,7 +15,7 @@ module.exports = (sequelize, dataTypes) => {
             allowNull: false
         },
         precio : {
-            type: dataTypes.FLOAT,
+            type: dataTypes.DECIMAL,
             allowNull: false
         },
         createdAt: {
